Revert "Refactor: project structure (#71)"
This reverts commit 85777998b433046fa5120ac9750df669ee103fe7.
