+++ app/api/helpers/string.py
