+++ app/api/models/base_manga.py
