+++ app/api/scrapers/base_manga.py
@@ -1,4 +1,4 @@
-from ..models.base_manga import MangaModel
+from app.api.models.base_manga import MangaModel
 from ..decorators.return_decorator import return_on_error
 from ..helpers.string import StringHelper
 from ..helpers.html_helper import HTMLHelper
