+++ app/api/scrapers/most_viewed.py
@@ -1,8 +1,8 @@
 from types import ClassMethodDescriptorType
 from selectolax.parser import Node
 
-from ..decorators.return_decorator import return_on_error
-from ..helpers.scraper import ScrapeHelper
+from app.api.decorators.return_decorator import return_on_error
+from app.api.helpers.scraper import ScrapeHelper
 from ..helpers.html_helper import HTMLHelper
