--- main.py
@@ -1,4 +1,0 @@
-import uvicorn
-
-if __name__ == "__main__":
-    uvicorn.run("server.api:app", host="0.0.0.0", port=8000, reload=True)
