+++ vercel.json
@@ -1,14 +1,14 @@
 {
-  	"builds": [
-	    {
-	    	"src": "/server/api.py",
-	    	"use": "@now/python"
-	    }
-  	],
-  	"routes": [
-	    {
-	    	"src": "/(.*)",
-	    	"dest": "server/api.py"
-	    }
-  	]
+  "builds": [
+    {
+      "src": "app/main.py",
+      "use": "@vercel/python"
+    }
+  ],
+  "routes": [
+    {
+      "src": "/(.*)",
+      "dest": "app/main.py"
+    }
+  ]
 }
